daos: dcp: abstract and improve daos code
Added an early check to make sure at least two
paths are provided, right after processing the arguments.
This will cut down on unnecessary processing.

daos_check_prefix now effectively ignores
the trailing '/' on the prefix and path.

If the daos prefix matches the source and source UUIDs are supplied,
alert the user and exit.

If the daos prefix matches the destination and destination UUIDs are
supplied, alert the user and exit.

If the daos source path is UNS and source UUIDs are supplied,
alert the user and exit.

If the daos destination path is UNS and destination UUIDs are supplied,
alert the user and exit.

DAOS-related code has been abstracted.
New daos_args struct to hold args.
New daos_setup to setup DAOS.
New daos_cleanup to cleanup DAOS.

Signed-off-by: Dalton Bohning <daltonx.bohning@intel.com>